Renames, remove conf
