+++ src/main/kotlin/com/github/vol0n/utbotcppclion/services/UTBotStartupActivity.kt
@@ -1,5 +1,6 @@
 package com.github.vol0n.utbotcppclion.services
 
+import com.github.vol0n.utbotcppclion.client.Client
 import com.intellij.openapi.components.service
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.startup.StartupActivity
